+++ frontend/pages/pokemons/index.tsx
@@ -1,84 +1,49 @@
-import {
-  Button,
-  Loading,
-  ProgressBar,
-  Search,
-  ContentSwitcher,
-  Switch,
-  Dropdown,
-  IconButton,
-} from '@carbon/react'
-import { Add, Edit } from '@carbon/icons-react'
+import { Header } from '@/components/pokemons/Header'
+import { Button, Loading, ProgressBar } from '@carbon/react'
+import { Add } from '@carbon/icons-react'
+import Head from 'next/head'
 import React from 'react'
-
-const items = [
-  {
-    id: 'option-0',
-    text: 'Lorem, ipsum dolor sit amet consectetur adipisicing elit.',
-  },
-  {
-    id: 'option-1',
-    text: 'Option 1',
-  },
-  {
-    id: 'option-2',
-    text: 'Option 2',
-  },
-  {
-    id: 'option-3',
-    text: 'Option 3 - a disabled item',
-    disabled: true,
-  },
-  {
-    id: 'option-4',
-    text: 'Option 4',
-  },
-  {
-    id: 'option-5',
-    text: 'Option 5',
-  },
-]
+import styles from '@/styles/pokemons.module.scss'
 
 export const PokemonsPage: React.FC = () => {
   return (
-    <div>
-      Hi there Dummy pokemons page
-      <Button isExpressive>Hi there</Button>
-      <Button kind="tertiary" isExpressive>
-        Hi there
-      </Button>
-      <IconButton label="123 label">
-        <Edit />
-      </IconButton>
-      <Button
-        renderIcon={Add}
-        iconDescription="Icon Description"
-        hasIconOnly
-        onClick={() => console.log('onClick')}
-      />
-      <ContentSwitcher onChange={() => {}}>
-        <Switch name="one" text="First section" />
-        <Switch name="two">Second section</Switch>
-        <Switch name="three" text="Third section" />
-      </ContentSwitcher>
-      <Loading description="Active loading indicator" withOverlay={false} />
-      <ProgressBar
-        label="Progress bar label"
-        helperText="Optional helper text"
-        value={75}
-      />
-      <Search />
-      <Dropdown
-        id="inline"
-        titleText="Inline dropdown label"
-        label="Dropdown menu options"
-        type="inline"
-        items={items}
-        itemToString={(item: { id: string; text: string }) =>
-          item ? item.text : ''
-        }
-      />
-    </div>
+    <>
+      <Head>
+        <title>Pokemons</title>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className={styles.main}>
+        <Header />
+        <div>
+          Content
+          <div>
+            Hi there Dummy pokemons page
+            <Button isExpressive>Hi there</Button>
+            <Button kind="tertiary" isExpressive>
+              Hi there
+            </Button>
+            <Button
+              renderIcon={Add}
+              iconDescription="Icon Description"
+              hasIconOnly
+              onClick={() => console.log('onClick')}
+            />
+            <Loading
+              description="Active loading indicator"
+              withOverlay={false}
+            />
+            <ProgressBar
+              label="Progress bar label"
+              helperText="Optional helper text"
+              value={75}
+            />
+          </div>
+        </div>
+      </main>
+    </>
   )
 }
